ci: prevent stack configs from overwriting eachother in tests
